+++ packages/website/lib/utils.js
@@ -29,7 +29,8 @@ export const formatTimestampFull = timestamp => {
  * @param {string} string The string being truncated
  * @param {number} [len] The max length allowed
  * @param {string} [end] The copy used to truncate at the end
- * @param {'single'|'double'} [type] The type of truncation to use, `single` puts ellipses at the end, `double` in the middle
+ * @param {'single'|'double'} [type] The type of truncation to use, `single` puts ellipses at the end,
+ *  `double` in the middle
  * @returns {string}
  */
 export const truncateString = (string, len = 30, end = '...', type = 'single') => {
@@ -45,19 +46,19 @@ export const truncateString = (string, len = 30, end = '...', type = 'single') =
  *
  * @param {string} text Text to be copied to clipboard
  */
-export const addTextToClipboard = (text) => {
-  const container = document.createElement('textarea')
-  container.style.position = 'fixed'
-  container.style.left = '-99999px'
-  container.style.zIndex = '-1'
-  container.style.opacity = '0'
-  container.style.pointerEvents = 'none'
-  container.innerHTML = text
-  document.body.appendChild(container)
-  container.select()
-  document.execCommand('copy')
-  document.body.removeChild(container)
-}
+export const addTextToClipboard = text => {
+  const container = document.createElement('textarea');
+  container.style.position = 'fixed';
+  container.style.left = '-99999px';
+  container.style.zIndex = '-1';
+  container.style.opacity = '0';
+  container.style.pointerEvents = 'none';
+  container.innerHTML = text;
+  document.body.appendChild(container);
+  container.select();
+  document.execCommand('copy');
+  document.body.removeChild(container);
+};
 
 /**
  * Utility function to standardize element heights based on largest sibling
@@ -67,29 +68,28 @@ export const addTextToClipboard = (text) => {
  */
 export const standardizeSiblingHeights = (target, reset) => {
   if (typeof document !== 'undefined') {
-    const elements = (
+    const elements =
       /** @type {HTMLCollectionOf<HTMLElement>} */
-      (document.getElementsByClassName(target))
-    );
-    const heights = []
+      (document.getElementsByClassName(target));
+    const heights = [];
 
     for (let i = 0; i < elements.length; i++) {
-      const el = /** @type {HTMLElement} */ (elements[i])
+      const el = /** @type {HTMLElement} */ (elements[i]);
       if (reset) {
-        el.style.minHeight = 'unset'
+        el.style.minHeight = 'unset';
       }
-      const rect = el.getBoundingClientRect()
-      heights.push(rect.height)
+      const rect = el.getBoundingClientRect();
+      heights.push(rect.height);
     }
 
     const max = Math.max(...heights);
 
     for (let i = 0; i < elements.length; i++) {
-      const el = /** @type {HTMLElement} */ (elements[i])
-      el.style.minHeight = max + 'px'
+      const el = /** @type {HTMLElement} */ (elements[i]);
+      el.style.minHeight = max + 'px';
     }
   }
-}
+};
 
 /**
  * Utility to check if element is in viewport
@@ -97,10 +97,10 @@ export const standardizeSiblingHeights = (target, reset) => {
  * @param {any} element element to test
  * @returns {boolean}
  */
-export const elementIsInViewport = (element) => {
+export const elementIsInViewport = element => {
   if (element && typeof window !== 'undefined' && typeof document !== 'undefined') {
     const rect = element.getBoundingClientRect();
-    return (rect.top >= 0 && rect.bottom <= (window.innerHeight || document.documentElement.clientHeight))
+    return rect.top >= 0 && rect.bottom <= (window.innerHeight || document.documentElement.clientHeight);
   }
-  return false
-}
+  return false;
+};